Major refactoring of survomatic core functionality. 1. A new function, fp() replaces findpars() which will be removed after the completion of testing. Instead of attempting to combine reporting and analysis, fp() generates a data object (of class 'fp'. This object can be captured and later passed to a dedicated print() function to generate human-readable results and a dedicated plot() function to generate various graphs of the data.
2. We have agreed that using simulations to find "empirical null distributions" for data is cicrular
reasoning and therefore in fp() models are fitted only using the chi-square distribution.
However, we still believe that permutations are a valid alternative where it comes to 
determining whether constrained models differ from unconstrained ones, and now that 
functionality is transparently built into fp()-- by default fp() permutes the two groups
being compared and generates empirical null distritutions.

3. fp() is much more modular and will be easier to maintain than findpars() and it is
better suited to inclusion in scripts.

4. fp() has replaced findpars() in survgui(). The output will hopefully be easier to interpret and
shows less irrelevant information.

5. The survgui() interface was cleaned up, most of the buttons have been moved to menus. The editable
tables have been replaced with "edit" buttons that take the user to a table editor that opens
in a separate window. The wierd behavior when the main window gets resized has been fixed.

6. survgui() makes available in its menus almost all the same graphs as can be obtained through the
plot.fp() function on the command-line.

7. The degree to which opsurv() spams your screen can now be controlled-- set verbose=0 to make it the
quietest, verbose=2 to make it show progress on the screen like the pre 1.4 versions did, and
verbose=1 as a middle ground between screen-spam and not knowing whether or not the scrpt has 
gotten stuck.



git-svn-id: svn://svn.r-forge.r-project.org/svnroot/rmodest@44 edb9625f-4e0d-4859-8d74-9fd3b1da38cb